Improved parsing.
The IBH header wasn't actually part of the packets we see, that was a
mistake. Instead there's another 'pdu' header (libnodave doesn't name
it), which has a different length depending on its version.

At this point we *should* be getting at the interesting bits
(read/write, response data, DB number, ...)