[doc] Basic documentation on profinet design.
This isn't much, but it did result in major changes in my understanding
of the protocol and the design of the lib.

Basically, profinet appears to be built on top of the ISO transport
layer (COTP) which is layered on top of TCP using TPKT.

That means we need to implement TPKT (trivial) and COTP (a little more
complicated). It also means we have a smaller unknown packet to tackle.