Pcap bugfix: Deal with runt packets
We did not account for the fact that its possible (and common) for the
ethernet packets to be longer than needed for the actual payload.
Ethernet packets have a minimum length which means that short packets
(like a TCP ACK) have trailing bytes, which are not part of the payload.