Rename project to s7comm
It appears that I've misunderstood the terminology sufficiently to
incorrectly name the project.

The protocol implemented here is not Profinet, but a Siemens proprietary
thing. The specific variant appears to be referred to as 'RFC1006' in
Siemens-speak, which is silly, because RFC1006 is TKIP (ISO transport
layer encapsulated in TCP).
It may simply be MPI over TCP (or MPI over ISO/COTP over TCP) but
that's not clear to me at all.

Call is s7comm, because it does need a name, and I'll be using it to
talk to a Siemens S7 PLC. It's also the name used by a Wireshark plugin
module which analyses this protocol.