[analysis] Link statically against the lib
The shared object of the s7 lib deliberately hides its internals,
exposing only the minimal number of functions required to actually use
it. This is intended to make maintaining the API easier.
It's not so great for the analysis code, which wants to play with the
internals of the lib (i.e. use ppkt structures and use only selected
protocol layers).