[ppkt] Allow ppkt_chain_size() on NULL
Allow users to call ppkt_chain_size() with a NULL ppkt. This doesn't
make much sense at first glance, but it's a usefulf shortcut for the
layers which accumulate data until sufficient bytes are present to
process a header.

It allows them to do while (ppkt_chain_size(queue) > ...) rather than
having to explicitly check for NULL as well.

Consider a NULL pointer to be a chain of lenght zero.