Move the TPKT port number to the tcp file
Remove all knowledge of the TPKT port number from the tpkt layer and
move it all into the TCP layer.

There's no real reason we need to support different port numbers.
RFC1006 specifies port 102 and it's unlikely we'll ever see a PLC using
a different port number.

The biggest advantage though is that it means the arguments for
tcp_connect() are now the same as for all the other protocol layers.