Split connect into open and connect
Split up the protocol method 'connect' into an open and connect function
(for all protocol layers). The idea is for open to create a protocol
object but not actually open the connection yet. connect then actually
sets up the connection (performing handshakes as required).

This means we can chose to only use open in the analysis code. Trying to
send packets to pcap (and waiting for responses) doesn't work very well.